Added configuration file and fixed all hard coded directories.

+++ BRSReadout/BRSReadout/Camera.cs
@@ -10,6 +10,7 @@ using System.Runtime.InteropServices;
 using System.Diagnostics;
 using System.IO;
 using System.Threading;
+using System.Configuration;
 using PylonC.NET;
 using TwinCAT.Ads;
 using BRSReadout;
@@ -20,12 +21,14 @@ using BRSReadout;
  */
 class Camera
 {
+    public string cameraType = ConfigurationManager.AppSettings.Get("camera");
+
     public delegate void frameCallbackDelegate(PylonBuffer<Byte> buffer);
     public delegate void dataDelegate(ushort[] data);
 
     dataDelegate masterDataDelegate;
     public ushort[] data = new ushort[4096];
-
+  
     PYLON_DEVICE_HANDLE hDev = new PYLON_DEVICE_HANDLE();
 
     uint numDevices;
@@ -43,8 +46,6 @@ class Camera
     string ip;
     uint j;
 
-
-
     public Camera()
     {
     }
@@ -53,57 +54,60 @@ class Camera
         //Initiates camera to run in 12 bit mode with continuous acquisition.
         int i = 0;
         uint NUM_BUFFERS = 10;
-        Pylon.Initialize();
-        numDevices = Pylon.EnumerateDevices();
-        while (ip != "10.110.136.138")
+        if (cameraType == "basler")
         {
-            hDev = Pylon.CreateDeviceByIndex(j);
-            prop = Pylon.DeviceGetDeviceInfoHandle(hDev);
-            ip = Pylon.DeviceInfoGetPropertyValueByIndex(prop, 8);
-            j++;
-            if (j > 100) { break; }
-        }
-        try
-        {
-            Pylon.DeviceOpen(hDev, Pylon.cPylonAccessModeControl | Pylon.cPylonAccessModeStream);
-        }
-        catch (Exception ex)
-        {
-            EmailError.emailAlert(ex);
-            throw (ex);
-        }
+            Pylon.Initialize();
+            numDevices = Pylon.EnumerateDevices();
+            while (ip != "10.110.136.138")
+            {
+                hDev = Pylon.CreateDeviceByIndex(j);
+                prop = Pylon.DeviceGetDeviceInfoHandle(hDev);
+                ip = Pylon.DeviceInfoGetPropertyValueByIndex(prop, 8);
+                j++;
+                if (j > numDevices) { break; }
+            }
+            try
+            {
+                Pylon.DeviceOpen(hDev, Pylon.cPylonAccessModeControl | Pylon.cPylonAccessModeStream);
+            }
+            catch (Exception ex)
+            {
+                EmailError.emailAlert(ex);
+                throw (ex);
+            }
 
-        isAvail = Pylon.DeviceFeatureIsAvailable(hDev, "EnumEntry_PixelFormat_Mono12");
-        Pylon.DeviceFeatureFromString(hDev, "PixelFormat", "Mono12");
-        Pylon.DeviceFeatureFromString(hDev, "AcquisitionMode", "Continuous");
-        Pylon.DeviceSetIntegerFeature(hDev, "Height", 1);
-        Pylon.DeviceSetFloatFeature(hDev, "ExposureTimeAbs", 500); //Exposure time is in microseconds and rounded to the closest 100 ns.
-        isAvail = Pylon.DeviceFeatureIsWritable(hDev, "GevSCPSPacketSize");
-        if (isAvail)
-        {
-            Pylon.DeviceSetIntegerFeature(hDev, "GevSCPSPacketSize", 1500);
+            isAvail = Pylon.DeviceFeatureIsAvailable(hDev, "EnumEntry_PixelFormat_Mono12");
+            Pylon.DeviceFeatureFromString(hDev, "PixelFormat", "Mono12");
+            Pylon.DeviceFeatureFromString(hDev, "AcquisitionMode", "Continuous");
+            Pylon.DeviceSetIntegerFeature(hDev, "Height", 1);
+            Pylon.DeviceSetFloatFeature(hDev, "ExposureTimeAbs", 500); //Exposure time is in microseconds and rounded to the closest 100 ns.
+            isAvail = Pylon.DeviceFeatureIsWritable(hDev, "GevSCPSPacketSize");
+            if (isAvail)
+            {
+                Pylon.DeviceSetIntegerFeature(hDev, "GevSCPSPacketSize", 1500);
+            }
+            payloadSize = checked((uint)Pylon.DeviceGetIntegerFeature(hDev, "PayloadSize"));
+            nStreams = Pylon.DeviceGetNumStreamGrabberChannels(hDev);
+            hGrabber = Pylon.DeviceGetStreamGrabber(hDev, 0);
+            Pylon.StreamGrabberOpen(hGrabber);
+            hWait = Pylon.StreamGrabberGetWaitObject(hGrabber);
+            Pylon.StreamGrabberSetMaxNumBuffer(hGrabber, NUM_BUFFERS);
+            Pylon.StreamGrabberSetMaxBufferSize(hGrabber, payloadSize);
+            Pylon.StreamGrabberPrepareGrab(hGrabber);
+            buffers = new Dictionary<PYLON_STREAMBUFFER_HANDLE, PylonBuffer<Byte>>();
+            for (i = 0; i < NUM_BUFFERS; ++i)
+            {
+                PylonBuffer<Byte> buffer = new PylonBuffer<byte>(payloadSize, true);
+                PYLON_STREAMBUFFER_HANDLE handle = Pylon.StreamGrabberRegisterBuffer(hGrabber, ref buffer);
+                buffers.Add(handle, buffer);
+            }
+            i = 0;
+            foreach (KeyValuePair<PYLON_STREAMBUFFER_HANDLE, PylonBuffer<Byte>> pair in buffers)
+            {
+                Pylon.StreamGrabberQueueBuffer(hGrabber, pair.Key, i++);
+            }
+            Pylon.DeviceExecuteCommandFeature(hDev, "AcquisitionStart");
         }
-        payloadSize = checked((uint)Pylon.DeviceGetIntegerFeature(hDev, "PayloadSize"));
-        nStreams = Pylon.DeviceGetNumStreamGrabberChannels(hDev);
-        hGrabber = Pylon.DeviceGetStreamGrabber(hDev, 0);
-        Pylon.StreamGrabberOpen(hGrabber);
-        hWait = Pylon.StreamGrabberGetWaitObject(hGrabber);
-        Pylon.StreamGrabberSetMaxNumBuffer(hGrabber, NUM_BUFFERS);
-        Pylon.StreamGrabberSetMaxBufferSize(hGrabber, payloadSize);
-        Pylon.StreamGrabberPrepareGrab(hGrabber);
-        buffers = new Dictionary<PYLON_STREAMBUFFER_HANDLE, PylonBuffer<Byte>>();
-        for (i = 0; i < NUM_BUFFERS; ++i)
-        {
-            PylonBuffer<Byte> buffer = new PylonBuffer<byte>(payloadSize, true);
-            PYLON_STREAMBUFFER_HANDLE handle = Pylon.StreamGrabberRegisterBuffer(hGrabber, ref buffer);
-            buffers.Add(handle, buffer);
-        }
-        i = 0;
-        foreach (KeyValuePair<PYLON_STREAMBUFFER_HANDLE, PylonBuffer<Byte>> pair in buffers)
-        {
-            Pylon.StreamGrabberQueueBuffer(hGrabber, pair.Key, i++);
-        }
-        Pylon.DeviceExecuteCommandFeature(hDev, "AcquisitionStart");
     }
 
     public void startFrameGrab(int nr, int trigmode, dataDelegate dd)
@@ -114,17 +118,20 @@ class Camera
         {
             try
             {
-                PylonBuffer<Byte> buffer;
-                Pylon.WaitObjectWait(hWait, 100);
+                if (cameraType == "basler")
+                {
+                    PylonBuffer<Byte> buffer;
+                    Pylon.WaitObjectWait(hWait, 100);
 
-                Pylon.StreamGrabberRetrieveResult(hGrabber, out grabResult);
+                    Pylon.StreamGrabberRetrieveResult(hGrabber, out grabResult);
 
-                bufferIndex = (int)grabResult.Context;
-                while (buffers.TryGetValue(grabResult.hBuffer, out buffer) != true) ;
-                masterDataDelegate = dd;
-                Buffer.BlockCopy(buffer.Array, 0, data, 0, buffer.Array.Length);
-                masterDataDelegate(data);
-                Pylon.StreamGrabberQueueBuffer(hGrabber, grabResult.hBuffer, bufferIndex);
+                    bufferIndex = (int)grabResult.Context;
+                    while (buffers.TryGetValue(grabResult.hBuffer, out buffer) != true) ;
+                    masterDataDelegate = dd;
+                    Buffer.BlockCopy(buffer.Array, 0, data, 0, buffer.Array.Length);
+                    masterDataDelegate(data);
+                    Pylon.StreamGrabberQueueBuffer(hGrabber, grabResult.hBuffer, bufferIndex);
+                }
             }
             catch (Exception ex)
             {
@@ -136,34 +143,37 @@ class Camera
 
     public void stopFrameGrab()
     {
-        //Stops acquisition and closes camera.
-        Pylon.DeviceExecuteCommandFeature(hDev, "AcquisitionStop");
-
-        Pylon.StreamGrabberCancelGrab(hGrabber);
-
-        do
+        if (cameraType == "basler")
         {
-            isReady = Pylon.StreamGrabberRetrieveResult(hGrabber, out grabResult);
+            //Stops acquisition and closes camera.
+            Pylon.DeviceExecuteCommandFeature(hDev, "AcquisitionStop");
 
-        } while (isReady);
+            Pylon.StreamGrabberCancelGrab(hGrabber);
 
-        foreach (KeyValuePair<PYLON_STREAMBUFFER_HANDLE, PylonBuffer<Byte>> pair in buffers)
-        {
-            Pylon.StreamGrabberDeregisterBuffer(hGrabber, pair.Key);
-            pair.Value.Dispose();
+            do
+            {
+                isReady = Pylon.StreamGrabberRetrieveResult(hGrabber, out grabResult);
+
+            } while (isReady);
+
+            foreach (KeyValuePair<PYLON_STREAMBUFFER_HANDLE, PylonBuffer<Byte>> pair in buffers)
+            {
+                Pylon.StreamGrabberDeregisterBuffer(hGrabber, pair.Key);
+                pair.Value.Dispose();
+            }
+            buffers = null;
+
+            Pylon.StreamGrabberFinishGrab(hGrabber);
+
+            Pylon.StreamGrabberClose(hGrabber);
+
+            Pylon.DeviceClose(hDev);
+            Pylon.DestroyDevice(hDev);
+
+            Console.ReadLine();
+
+            Pylon.Terminate();
         }
-        buffers = null;
-
-        Pylon.StreamGrabberFinishGrab(hGrabber);
-
-        Pylon.StreamGrabberClose(hGrabber);
-
-        Pylon.DeviceClose(hDev);
-        Pylon.DestroyDevice(hDev);
-
-        Console.ReadLine();
-
-        Pylon.Terminate();
 
     }
